+++ Sources/MacIMEKit/IME.swift
@@ -98,7 +98,16 @@ public enum IME {
 
       let src = try select(id: prev_id)
 
-      if state.cjkRefresh { CJK.refresh() }
+      if state.cjkRefresh {
+         if Thread.isMainThread {
+            CJK.refresh()
+         } else {
+            // DispatchQueue.main.sync { // WORKS
+            DispatchQueue.main.async {
+               CJK.refresh()
+            }
+         }
+      }
 
       return src.id
    }