+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [4.0.2](https://github.com/riodelphino/macime/compare/v4.0.1...v4.0.2) (2026-02-16)
 
+* **fix:** Restrict NSApp to work only with `CJK.swift`
 * **fix:** Activate CJK window 
 
 ## [4.0.1](https://github.com/riodelphino/macime/compare/v4.0.0...v4.0.1) (2026-02-14)